Fix Qt6 deprecation warnings in QVariant and range-for loops
- Replace QVariant::convert(int) with QVariant::convert(QMetaType)
- Replace QVariant::type() with typeId() in fileio and translatemanager
- Fix range-for loop variable type mismatch (long -> int) in LOV dumpers
- Fix signed/unsigned comparison in allsky property loop

+++ libs/utils/elementjsondumper.cpp
@@ -81,7 +81,7 @@ void ElementJsonDumper::visit(ElementInt *pElement, QVariantMap &data, bool &emi
         if (pElement->getLov().size() > 0)
         {
             QJsonObject lines = QJsonObject();
-            for(const long &key : pElement->getLov().keys()) {
+            for(const int &key : pElement->getLov().keys()) {
                 lines[QString::number(key)] = pElement->getLov()[key];
             }
             json["listOfValues"] = lines;
